requirements.txt
pip
discord.py